WIN32 - started working in raw input mouse data
This process bypasses the default windows input events to increase
performance.

+++ src/ccore/core/common/interface/event.h
@@ -145,6 +145,6 @@ typedef struct {
 		ccMouseButtonType mouseButton;
 		ccPoint mousePosition;
 		ccKeyCode key;
-		int scrollDelta;
+		short scrollDelta;
 	};
 } ccEvent;
